+++ bankroll/__init__.py
@@ -2,4 +2,4 @@ from .aggregator import DataAggregator
 from .analysis import *
 from .brokers import *
 from .model import *
-from .etf import *
+from .portfolio import *
